// base/free5gc/webconsole/frontend/src/api/api.ts
export interface AccessAndMobilitySubscriptionData {
    /**
     * 
     * @type {Array<string>}
     * @memberof AccessAndMobilitySubscriptionData
     */
    'gpsis'?: Array<string>;
    /**
     * 
     * @type {SubscribedUeAmbr}
     * @memberof AccessAndMobilitySubscriptionData
     */
    'subscribedUeAmbr'?: SubscribedUeAmbr;
    /**
     * 
     * @type {DefaultSingleNssais}
     * @memberof AccessAndMobilitySubscriptionData
     */
    'nssai'?: DefaultSingleNssais;
}
/**
 * 
 * @export
 * @interface AmPolicyData
 */
export interface AmPolicyData {
    /**
     * 
     * @type {Array<string>}
     * @memberof AmPolicyData
     */
    'subscCats'?: Array<string>;
}
/**
 * 
 * @export
 * @interface Arp
 */
export interface Arp {
    /**
     * 
     * @type {number}
     * @memberof Arp
     */
    'priorityLevel': number;
    /**
     * 
     * @type {string}
     * @memberof Arp
     */
    'preemptCap': string;
    /**
     * 
     * @type {string}
     * @memberof Arp
     */
    'preemptVuln': string;
}
// ...
export interface ChargingData {
    /**
     * 
     * @type {string}
     * @memberof ChargingData
     */
    'snssai'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChargingData
     */
    'dnn': string;
    /**
     * 
     * @type {number}
     * @memberof ChargingData
     */
    'qosRef'?: number;
    /**
     * 
     * @type {string}
     * @memberof ChargingData
     */
    'filter': string;
    /**
     * 
     * @type {string}
     * @memberof ChargingData
     */
    'chargingMethod'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChargingData
     */
    'quota'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChargingData
     */
    'unitCost'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChargingData
     */
    'ueId'?: string;
}
/**
 * 
 * @export
 * @interface DefaultSingleNssais
 */
export interface DefaultSingleNssais {
    /**
     * 
     * @type {Array<Nssai>}
     * @memberof DefaultSingleNssais
     */
    'defaultSingleNssais': Array<Nssai>;
    /**
     * 
     * @type {Array<Nssai>}
     * @memberof DefaultSingleNssais
     */
    'singleNssais'?: Array<Nssai>;
}
/**
 * 
 * @export
 * @interface Dnn
 */
export interface Dnn {
    /**
     * 
     * @type {string}
     * @memberof Dnn
     */
    'dnn': string;
}
/**
 * 
 * @export
 * @interface DnnConfiguration
 */
export interface DnnConfiguration {
    /**
     * 
     * @type {PduSessionTypes}
     * @memberof DnnConfiguration
     */
    'pduSessionTypes': PduSessionTypes;
    /**
     * 
     * @type {SscModes}
     * @memberof DnnConfiguration
     */
    'sscModes': SscModes;
    /**
     * 
     * @type {Model5gQosProfile}
     * @memberof DnnConfiguration
     */
    '5gQosProfile'?: Model5gQosProfile;
    /**
     * 
     * @type {SessionAmbr}
     * @memberof DnnConfiguration
     */
    'sessionAmbr'?: SessionAmbr;
    /**
     * 
     * @type {Array<IpAddress>}
     * @memberof DnnConfiguration
     */
    'staticIpAddress'?: Array<IpAddress>;
    /**
     * 
     * @type {UpSecurity}
     * @memberof DnnConfiguration
     */
    'upSecurity'?: UpSecurity;
}
// ...
export interface FlowRules {
    /**
     * 
     * @type {string}
     * @memberof FlowRules
     */
    'filter'?: string;
    /**
     * 
     * @type {number}
     * @memberof FlowRules
     */
    'precedence'?: number;
    /**
     * 
     * @type {string}
     * @memberof FlowRules
     */
    'snssai'?: string;
    /**
     * 
     * @type {string}
     * @memberof FlowRules
     */
    'dnn'?: string;
    /**
     * 
     * @type {number}
     * @memberof FlowRules
     */
    'qosRef'?: number;
}
/**
 * 
 * @export
 * @interface IpAddress
 */
export interface IpAddress {
    /**
     * 
     * @type {string}
     * @memberof IpAddress
     */
    'ipv4Addr'?: string;
    /**
     * 
     * @type {string}
     * @memberof IpAddress
     */
    'ipv6Addr'?: string;
    /**
     * 
     * @type {string}
     * @memberof IpAddress
     */
    'ipv6Prefix'?: string;
}
// ...
export interface Model5gQosProfile {
    /**
     * 
     * @type {number}
     * @memberof Model5gQosProfile
     */
    '5qi': number;
    /**
     * 
     * @type {Arp}
     * @memberof Model5gQosProfile
     */
    'arp': Arp;
    /**
     * 
     * @type {number}
     * @memberof Model5gQosProfile
     */
    'priorityLevel'?: number;
}
/**
 * 
 * @export
 * @interface Nssai
 */
export interface Nssai {
    /**
     * 
     * @type {number}
     * @memberof Nssai
     */
    'sst': number;
    /**
     * 
     * @type {string}
     * @memberof Nssai
     */
    'sd'?: string;
}
// ...
export interface PduSessionTypes {
    /**
     * 
     * @type {string}
     * @memberof PduSessionTypes
     */
    'defaultSessionType': string;
    /**
     * 
     * @type {Array<string>}
     * @memberof PduSessionTypes
     */
    'allowedSessionTypes'?: Array<string>;
}
// ...
export interface Profile {
    /**
     * 
     * @type {string}
     * @memberof Profile
     */
    'profileName': string;
    /**
     * 
     * @type {AccessAndMobilitySubscriptionData}
     * @memberof Profile
     */
    'AccessAndMobilitySubscriptionData': AccessAndMobilitySubscriptionData;
    /**
     * 
     * @type {Array<SessionManagementSubscriptionData>}
     * @memberof Profile
     */
    'SessionManagementSubscriptionData': Array<SessionManagementSubscriptionData>;
    /**
     * 
     * @type {SmfSelectionSubscriptionData}
     * @memberof Profile
     */
    'SmfSelectionSubscriptionData': SmfSelectionSubscriptionData;
    /**
     * 
     * @type {AmPolicyData}
     * @memberof Profile
     */
    'AmPolicyData': AmPolicyData;
    /**
     * 
     * @type {SmPolicyData}
     * @memberof Profile
     */
    'SmPolicyData': SmPolicyData;
    /**
     * 
     * @type {Array<FlowRules>}
     * @memberof Profile
     */
    'FlowRules': Array<FlowRules>;
    /**
     * 
     * @type {Array<QosFlows>}
     * @memberof Profile
     */
    'QosFlows': Array<QosFlows>;
    /**
     * 
     * @type {Array<ChargingData>}
     * @memberof Profile
     */
    'ChargingDatas': Array<ChargingData>;
}
/**
 * 
 * @export
 * @interface ProfileListIE
 */
export interface ProfileListIE {
    /**
     * 
     * @type {string}
     * @memberof ProfileListIE
     */
    'profileName'?: string;
    /**
     * 
     * @type {string}
     * @memberof ProfileListIE
     */
    'gpsi'?: string;
}
/**
 * 
 * @export
 * @interface QosFlows
 */
export interface QosFlows {
    /**
     * 
     * @type {string}
     * @memberof QosFlows
     */
    'snssai': string;
    /**
     * 
     * @type {string}
     * @memberof QosFlows
     */
    'dnn': string;
    /**
     * 
     * @type {number}
     * @memberof QosFlows
     */
    'qosRef': number;
    /**
     * 
     * @type {number}
     * @memberof QosFlows
     */
    '5qi': number;
    /**
     * 
     * @type {string}
     * @memberof QosFlows
     */
    'mbrUL'?: string;
    /**
     * 
     * @type {string}
     * @memberof QosFlows
     */
    'mbrDL'?: string;
    /**
     * 
     * @type {string}
     * @memberof QosFlows
     */
    'gbrUL'?: string;
    /**
     * 
     * @type {string}
     * @memberof QosFlows
     */
    'gbrDL'?: string;
}
/**
 * 
 * @export
 * @interface SessionAmbr
 */
export interface SessionAmbr {
    /**
     * 
     * @type {string}
     * @memberof SessionAmbr
     */
    'uplink': string;
    /**
     * 
     * @type {string}
     * @memberof SessionAmbr
     */
    'downlink': string;
}
/**
 * 
 * @export
 * @interface SessionManagementSubscriptionData
 */
export interface SessionManagementSubscriptionData {
    /**
     * 
     * @type {Nssai}
     * @memberof SessionManagementSubscriptionData
     */
    'singleNssai': Nssai;
    /**
     * 
     * @type {{ [key: string]: DnnConfiguration; }}
     * @memberof SessionManagementSubscriptionData
     */
    'dnnConfigurations'?: { [key: string]: DnnConfiguration; };
}
/**
 * 
 * @export
 * @interface SessionRule
 */
export interface SessionRule {
    /**
     * 
     * @type {string}
     * @memberof SessionRule
     */
    'sessRuleId': string;
}
/**
 * 
 * @export
 * @interface SmPolicyData
 */
export interface SmPolicyData {
    /**
     * 
     * @type {{ [key: string]: SmPolicySnssai; }}
     * @memberof SmPolicyData
     */
    'smPolicySnssaiData': { [key: string]: SmPolicySnssai; };
}
/**
 * 
 * @export
 * @interface SmPolicySnssai
 */
export interface SmPolicySnssai {
    /**
     * 
     * @type {Nssai}
     * @memberof SmPolicySnssai
     */
    'snssai': Nssai;
    /**
     * 
     * @type {{ [key: string]: Dnn; }}
     * @memberof SmPolicySnssai
     */
    'smPolicyDnnData'?: { [key: string]: Dnn; };
}
/**
 * 
 * @export
 * @interface SmfSelectionSubscriptionData
 */
export interface SmfSelectionSubscriptionData {
    /**
     * 
     * @type {{ [key: string]: SubscribedSnssaiInfo; }}
     * @memberof SmfSelectionSubscriptionData
     */
    'subscribedSnssaiInfos'?: { [key: string]: SubscribedSnssaiInfo; };
}
/**
 * 
 * @export
 * @interface SscModes
 */
export interface SscModes {
    /**
     * 
     * @type {string}
     * @memberof SscModes
     */
    'defaultSscMode': string;
    /**
     * 
     * @type {Array<string>}
     * @memberof SscModes
     */
    'allowedSscModes'?: Array<string>;
}
/**
 * 
 * @export
 * @interface SubscribedSnssaiInfo
 */
export interface SubscribedSnssaiInfo {
    /**
     * 
     * @type {Array<Dnn>}
     * @memberof SubscribedSnssaiInfo
     */
    'dnnInfos': Array<Dnn>;
}
/**
 * 
 * @export
 * @interface SubscribedUeAmbr
 */
export interface SubscribedUeAmbr {
    /**
     * 
     * @type {string}
     * @memberof SubscribedUeAmbr
     */
    'uplink': string;
    /**
     * 
     * @type {string}
     * @memberof SubscribedUeAmbr
     */
    'downlink': string;
}
// ...
export interface UpSecurity {
    /**
     * 
     * @type {string}
     * @memberof UpSecurity
     */
    'upIntegr': string;
    /**
     * 
     * @type {string}
     * @memberof UpSecurity
     */
    'upConfid': string;
}